+++ tower-http/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- None.
+- Add `NormalizePath` middleware
 
 ## Changed
 
@@ -22,8 +22,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Fixed
 
 - **trace:** Correctly identify gRPC requests in default `on_response` callback ([#278])
+- **cors:** Panic if a wildcard (`*`) is passed to `AllowOrigin::list`. Use
+  `AllowOrigin::any()` instead ([#285])
 
 [#278]: https://github.com/tower-rs/tower-http/pull/278
+[#285]: https://github.com/tower-rs/tower-http/pull/285
 
 # 0.3.4 (June 06, 2022)